// web/packages/admin/src/types/api.ts
export interface NoticeAction {
  label: string
  method: 'GET' | 'POST'
  url: string
  style?: 'primary' | 'default' | 'error' | null
  confirm?: boolean
  comment?: boolean
}